fix(nextcloud)!: Move RichObjectParameter from spreed to core and fix rich parameter typing for broken apps
