+++ packages/neon_framework/packages/talk_app/pubspec.yaml
@@ -31,6 +31,7 @@ dependencies:
 
 dev_dependencies:
   build_runner: ^2.4.12
+  built_value: ^8.9.2
   custom_lint: ^0.6.5
   flutter_keyboard_visibility: ^6.0.0
   flutter_test:
