+++ packages/nextcloud/lib/src/api/core/patches/1-compatibility-29.json
