+++ packages/nextcloud/lib/src/api/core/patches/2-compatibility-28.json
