+++ packages/nextcloud/lib/src/api/notifications/notifications.openapi.json
@@ -112,7 +112,10 @@
                     "subjectRichParameters": {
                         "type": "object",
                         "additionalProperties": {
-                            "$ref": "#/components/schemas/RichObjectParameter"
+                            "type": "object",
+                            "additionalProperties": {
+                                "type": "object"
+                            }
                         }
                     },
                     "messageRich": {
@@ -121,7 +124,10 @@
                     "messageRichParameters": {
                         "type": "object",
                         "additionalProperties": {
-                            "$ref": "#/components/schemas/RichObjectParameter"
+                            "type": "object",
+                            "additionalProperties": {
+                                "type": "object"
+                            }
                         }
                     },
                     "icon": {
@@ -194,114 +200,6 @@
                         "type": "string"
                     },
                     "signature": {
-                        "type": "string"
-                    }
-                }
-            },
-            "RichObjectParameter": {
-                "type": "object",
-                "required": [
-                    "type",
-                    "id",
-                    "name"
-                ],
-                "properties": {
-                    "type": {
-                        "type": "string"
-                    },
-                    "id": {
-                        "type": "string"
-                    },
-                    "name": {
-                        "type": "string"
-                    },
-                    "server": {
-                        "type": "string"
-                    },
-                    "link": {
-                        "type": "string"
-                    },
-                    "call-type": {
-                        "type": "string",
-                        "enum": [
-                            "one2one",
-                            "group",
-                            "public"
-                        ]
-                    },
-                    "icon-url": {
-                        "type": "string"
-                    },
-                    "message-id": {
-                        "type": "string"
-                    },
-                    "boardname": {
-                        "type": "string"
-                    },
-                    "stackname": {
-                        "type": "string"
-                    },
-                    "size": {
-                        "type": "string"
-                    },
-                    "path": {
-                        "type": "string"
-                    },
-                    "mimetype": {
-                        "type": "string"
-                    },
-                    "preview-available": {
-                        "type": "string",
-                        "enum": [
-                            "yes",
-                            "no"
-                        ]
-                    },
-                    "mtime": {
-                        "type": "string"
-                    },
-                    "latitude": {
-                        "type": "string"
-                    },
-                    "longitude": {
-                        "type": "string"
-                    },
-                    "description": {
-                        "type": "string"
-                    },
-                    "thumb": {
-                        "type": "string"
-                    },
-                    "website": {
-                        "type": "string"
-                    },
-                    "visibility": {
-                        "type": "string",
-                        "enum": [
-                            "0",
-                            "1"
-                        ]
-                    },
-                    "assignable": {
-                        "type": "string",
-                        "enum": [
-                            "0",
-                            "1"
-                        ]
-                    },
-                    "conversation": {
-                        "type": "string"
-                    },
-                    "etag": {
-                        "type": "string"
-                    },
-                    "permissions": {
-                        "type": "string"
-                    },
-                    "width": {
-                        "type": "string"
-                    },
-                    "height": {
                         "type": "string"
                     }
                 }
@@ -1231,7 +1129,10 @@
                                         "default": {},
                                         "description": "Rich objects to fill the subject placeholders, {@see \\OCP\\RichObjectStrings\\Definitions}",
                                         "additionalProperties": {
-                                            "$ref": "#/components/schemas/RichObjectParameter"
+                                            "type": "object",
+                                            "additionalProperties": {
+                                                "type": "object"
+                                            }
                                         }
                                     },
                                     "messageParameters": {
@@ -1239,7 +1140,10 @@
                                         "default": {},
                                         "description": "Rich objects to fill the message placeholders, {@see \\OCP\\RichObjectStrings\\Definitions}",
                                         "additionalProperties": {
-                                            "$ref": "#/components/schemas/RichObjectParameter"
+                                            "type": "object",
+                                            "additionalProperties": {
+                                                "type": "object"
+                                            }
                                         }
                                     }
                                 }
