+++ packages/nextcloud/lib/src/api/notifications/patches/0-rich-parameters.json
@@ -1,0 +1,46 @@
+[
+  {
+    "op": "remove",
+    "path": "/components/schemas/RichObjectParameter"
+  },
+  {
+    "op": "replace",
+    "path": "/components/schemas/Notification/properties/subjectRichParameters/additionalProperties",
+    "value": {
+      "type": "object",
+      "additionalProperties": {
+        "type": "object"
+      }
+    }
+  },
+  {
+    "op": "replace",
+    "path": "/components/schemas/Notification/properties/messageRichParameters/additionalProperties",
+    "value": {
+      "type": "object",
+      "additionalProperties": {
+        "type": "object"
+      }
+    }
+  },
+  {
+    "op": "replace",
+    "path": "/paths/~1ocs~1v2.php~1apps~1notifications~1api~1{apiVersion3}~1admin_notifications~1{userId}/post/requestBody/content/application~1json/schema/properties/subjectParameters/additionalProperties",
+    "value": {
+      "type": "object",
+      "additionalProperties": {
+        "type": "object"
+      }
+    }
+  },
+  {
+    "op": "replace",
+    "path": "/paths/~1ocs~1v2.php~1apps~1notifications~1api~1{apiVersion3}~1admin_notifications~1{userId}/post/requestBody/content/application~1json/schema/properties/messageParameters/additionalProperties",
+    "value": {
+      "type": "object",
+      "additionalProperties": {
+        "type": "object"
+      }
+    }
+  }
+]
