+++ packages/nextcloud/lib/src/api/spreed/patches/1-rich-parameters.json
@@ -1,0 +1,19 @@
+[
+  {
+    "op": "remove",
+    "path": "/components/schemas/RichObjectParameter"
+  },
+  {
+    "op": "replace",
+    "path": "/components/schemas/BaseMessage/properties/messageParameters",
+    "value": {
+      "type": "object",
+      "additionalProperties": {
+        "type": "object",
+        "additionalProperties": {
+          "type": "object"
+        }
+      }
+    }
+  }
+]
